+++ internal/zone_loader/src/runtime_object.cpp
@@ -4,12 +4,12 @@ using namespace warpgate;
 
 RuntimeObject::RuntimeObject(std::span<uint8_t> subspan, uint32_t version): buf_(subspan) {
     actor_file_ = std::string((char*)buf_.data());
-    uint32_t offset = sizeof(float) + sizeof(uint32_t) + actor_file_.size() + 1;
+    uint32_t offset = (uint32_t)(sizeof(float) + sizeof(uint32_t) + actor_file_.size() + 1);
     uint32_t instance_count = this->instance_count();
     for(uint32_t i = 0; i < instance_count; i++) {
         Instance instance(buf_.subspan(offset), version);
         instances.push_back(instance);
-        offset += instance.size();
+        offset += (uint32_t)instance.size();
     }
     buf_ = buf_.first(offset);
 }
